Time: 0 ms (100.00%) | Memory: 8.7 MB (3.90%) - LeetSync

+++ 3192-maximum-xor-product/maximum-xor-product.cpp
@@ -8,7 +8,7 @@ public:
         ll xXorb = 0;
 
         // 49th bit to nth bit
-        for (ll i = 49; i >= n; i--) {
+        for (int i = 49; i >= n; i--) {
 
             bool a_ith_bit = (a &(1ll<<i)) > 0;
             bool b_ith_bit = (b &(1ll<<i)) > 0;
